Restructure files, add band reducer, action and thunk files
