+++ src/containers/App/index.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
-import './CSS/App.css';
-import Header from './Header.js'
-import BandInfo from './BandInfo.js'
-import { ebmBands } from './ebmDatabase.js'
+import './App.css';
+import Header from '../../components/Header'
+import BandInfo from '../BandInfo'
 
 class App extends Component {
   constructor() {
